After lift index first draft should be functional.

+++ src/main/scala/org/broadinstitute/hail/Utils.scala
@@ -567,7 +567,7 @@ class RichBoolean(val b: Boolean) extends AnyVal {
 
   def iff(that: Boolean): Boolean = b == that
 
-  def toInt: Double = if (b) 1 else 0
+  def toInt: Int = if (b) 1 else 0
 
   def toDouble: Double = if (b) 1d else 0d
 }
